fix a bit

+++ modules/asdf/init.fish
@@ -1,3 +1,3 @@
-if test -f ~/opt/asdf/completions/asdf.fish
+if test -f ~/opt/asdf/completions/asdf.fish -a ! -f ~/.config/fish/completions/asdf.fish
     ln -s ~/opt/asdf/completions/asdf.fish ~/.config/fish/completions
 end
